fix(build): satisfy new-nightly clippy/rustc lints across the workspace
The floating nightly (1.100.0-nightly cea272fa3 2026-09-07) ships
new and promoted lints that turn previously-clean code into errors
under `-D warnings`. All mechanical fixes, no behavior change:

- nestrs-mcp: `Option::map_or(true, ..)` -> `is_none_or(..)`
  (server.rs, surfaces.rs); `#[allow(dead_code)]` on the documented
  `subscription_sinks_for` seam — reserved for the resource-write
  fan-out path, unit-tested, no production caller wires it yet.
- nestrs-storage: drop needless `as u64` casts in the s3/gcs/azure
  backends (ObjectMeta::size is already u64).
- nestrs: remove redundant `return` in match arms and brace the
  cfg(not(authz)) tail (authn.rs); `map_err(|e| f(e))` ->
  `map_err(f)` (crud_macro.rs); `x == false` -> `!x` (policies test);
  `map_or(false, |v| v == expected)` -> `== Some(expected)` and
  `match` -> `matches!` (predicates.rs); needless `&` borrows
  (throttler.rs).
- nestrs/src/security: route_metadata_csv's only consumer is the
  authz-gated `policies` module; export it under
  #[cfg(feature = "authz")] so default (featureless) builds don't
  see dead code. The public API under --all-features — what the
  snapshot pins — is unchanged.
- nestrs-microservices: redact_url is only called by the optional
  nats/redis/rabbitmq transports; cfg_attr-allow it in lib builds
  with none of those features (its #[cfg(test)] redaction tests
  still run in every config).
- tests: `(n + 2) / 3` -> `n.div_ceil(3)` (row_level_http); gate the
  cache-redis-only `ThrottlerBackend` trait import behind
  #[cfg(feature = "cache-redis")] (throttler).

Audit ref: production/security audit 2026-09-10 (workspace), nightly
lint sweep follow-up (Info)

+++ nestrs-storage/src/cloud_backends.rs
@@ -191,7 +191,7 @@ mod s3 {
                 let r = item.map_err(map_err)?;
                 out.push(ObjectMeta {
                     key: path_to_key(&r.location),
-                    size: r.size as u64,
+                    size: r.size,
                     last_modified_ms: Some(r.last_modified.timestamp_millis()),
                     content_type: None,
                     etag: r.e_tag,
@@ -333,7 +333,7 @@ mod gcs {
                 let r = item.map_err(map_err)?;
                 out.push(ObjectMeta {
                     key: path_to_key(&r.location),
-                    size: r.size as u64,
+                    size: r.size,
                     last_modified_ms: Some(r.last_modified.timestamp_millis()),
                     content_type: None,
                     etag: r.e_tag,
@@ -484,7 +484,7 @@ mod azure {
                 let r = item.map_err(map_err)?;
                 out.push(ObjectMeta {
                     key: path_to_key(&r.location),
-                    size: r.size as u64,
+                    size: r.size,
                     last_modified_ms: Some(r.last_modified.timestamp_millis()),
                     content_type: None,
                     etag: r.e_tag,
